code_base/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
pillow==10.1.0
razorpay==1.4.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.13.1
pytest==8.0.0
httpx==0.26.0
prometheus-client==0.19.0
sentry-sdk==1.39.1
python-json-logger==2.0.7 